Remove built exe from Sources since it is too large
The exe is already inside the com.elgato.youtube-dl-plugin.streamDeckPlugin file, so no need to commit it again separately, so might as well remove it since uncompressed it is too large and requires gitlfs. Devs will need to rebuild the exe from the visual studio project if they want to repackage a new streamDeckPlugin file.